Always raise lockPick event
Mirror the previous changes to trapDisarm, allowing lockPick to be
raised even on an unlocked door or container.  Unfortunately the
function is never called by Morrowind in that case, but it is left
open to be patched later.

Added event data item "lockPresent" to lockPick - true if a lock is
present; false otherwise.  As noted above, currently always true.

Added output event data item "clearTarget" to lockPick - set this to
clear the player's target as would otherwise happen on successfully
picking a lock.