Remove extra calls to math.random().
These were originally included because the standard library rand()
implementation is terrible, but LuaJIT does not use it.

From https://luajit.org/extensions.html
LuaJIT uses a Tausworthe PRNG with period 2^223 to implement
math.random() and math.randomseed(). The quality of the PRNG results is
much superior compared to the standard Lua implementation which uses the
 platform-specific ANSI rand().