Experimental changes to Reference::disable()
This is confirmed to be a lot more stable than the previous implementation and the only issue seem to be summons when expiring (for obvious reasons). Disabled bit is being set at the end as its interrupts enterLeaveSimulation logic. The call to TES3_MobileActor_onDeath is a bit overkill, but it contains a lot of functions that would be called anyways. Will be cleaned up when the exact needed behaviour has been determined. The main components being removal of actors from projectiles (missing), magic projectiles and magic sources (MobileActor::retireMagic), as well as friendly/hostile lists (MACT::sub_51FEB0, iterated over listAllActors in the worldController). The missing step seems to be removal from non-magic projectiles, which would only be an issue for summons or extreme edge-cases during normal gameplay. Since magic projectiles do get destroyed, I think destroying regular projectiles is a must in this case. Proper implementation of that will follow later.