Added cleanup of mobile actors magic effects when disabling references
After some testing this seems to be the only necessary functionality from the steps taken in TES3_MobileActor__onDeath. Cleanup from other actors data and lists is taken care of in the destructor of the mobile actor. Cleanup of non-magic projectiles will be handled in the mobile deactivated event. EnterLeaveSimulation is already executed in removeMob.